requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.18.4
antlr4-python3-runtime==4.9.3
asteroid-filterbanks==0.4.0
async-timeout==5.0.1
attrs==26.1.0
av==17.0.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
coloredlogs==15.0.1
colorlog==6.10.1
contourpy==1.3.2
ctranslate2==4.7.1
cycler==0.12.1
dotenv==0.9.9
einops==0.8.2
faster-whisper==1.2.1
filelock==3.25.2
flatbuffers==25.12.19
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.2.0
googleapis-common-protos==1.73.0
grpcio==1.78.0
hf-xet==1.4.2
huggingface_hub==0.36.2
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.1
Jinja2==3.1.6
joblib==1.5.3
julius==0.2.7
kiwisolver==1.5.0
lightning==2.6.1
lightning-utilities==0.15.3
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
networkx==3.4.2
nltk==3.9.3
numpy==2.2.6
omegaconf==2.3.0
onnxruntime==1.23.2
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-exporter-otlp-proto-http==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
optuna==4.8.0
pandas==2.3.3
pillow==12.1.1
primePy==1.3
propcache==0.4.1
protobuf==6.33.6
pyannote-audio==4.0.4
pyannote-core==6.0.1
pyannote-database==6.1.1
pyannote-metrics==4.0.0
pyannote-pipeline==4.0.0
pyannoteai-sdk==0.4.0
pydub==0.25.1
pyparsing==3.3.2
python-dotenv==1.2.2
pytorch-lightning==2.6.1
pytorch-metric-learning==2.9.0
pytz==2026.1.post1
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
sortedcontainers==2.4.0
SQLAlchemy==2.0.48
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
tomli==2.4.0
torch==2.8.0
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.8.0
torchcodec==0.10.0
torchmetrics==1.9.0
tqdm==4.67.3
transformers==4.57.6
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
whisperx==3.8.2
yarl==1.23.0
zipp==3.23.0